+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-Fixed a error when launching the app by opening a log on macOS through Finder.
+Minor usability improvements:
+* Added separate "Connect to Robot" and "Connect to Simulator" buttons in menu.
+* Moved all default log tables to the end of the list in the side bar.